+++ src/test/java/uk/gov/pay/connector/it/base/ChargingITestBase.java
@@ -12,13 +12,13 @@ import org.junit.After;
 import org.junit.Before;
 import uk.gov.pay.connector.charge.model.ServicePaymentReference;
 import uk.gov.pay.connector.charge.model.domain.ChargeStatus;
-import uk.gov.pay.connector.rules.CardidStub;
-import uk.gov.pay.connector.rules.LedgerStub;
 import uk.gov.pay.connector.it.dao.DatabaseFixtures;
 import uk.gov.pay.connector.junit.DropwizardTestContext;
 import uk.gov.pay.connector.junit.TestContext;
 import uk.gov.pay.connector.model.domain.AuthCardDetailsFixture;
+import uk.gov.pay.connector.rules.CardidStub;
 import uk.gov.pay.connector.rules.EpdqMockClient;
+import uk.gov.pay.connector.rules.LedgerStub;
 import uk.gov.pay.connector.rules.SmartpayMockClient;
 import uk.gov.pay.connector.rules.WorldpayMockClient;
 import uk.gov.pay.connector.util.AddGatewayAccountCredentialsParams;
@@ -72,8 +72,24 @@ public class ChargingITestBase {
     protected static final String RETURN_URL = "http://service.local/success-page/";
     protected static final String EMAIL = randomAlphabetic(242) + "@example.com";
     protected static final long AMOUNT = 6234L;
+    protected static final String JSON_REFERENCE_VALUE = "Test reference";
+    protected static final String JSON_DESCRIPTION_VALUE = "Test description";
 
     protected static final String SERVICE_ID = "external-service-id";
+    protected static final String JSON_PROVIDER_KEY = "payment_provider";
+    protected static final String PROVIDER_NAME = "sandbox";
+    protected static final String JSON_CHARGE_KEY = "charge_id";
+    protected static final String JSON_MESSAGE_KEY = "message";
+    protected static final String JSON_AMOUNT_KEY = "amount";
+    protected static final String JSON_REFERENCE_KEY = "reference";
+    protected static final String JSON_DESCRIPTION_KEY = "description";
+    protected static final String JSON_RETURN_URL_KEY = "return_url";
+    protected static final String JSON_LANGUAGE_KEY = "language";
+    protected static final String JSON_EMAIL_KEY = "email";
+    protected static final String JSON_MOTO_KEY = "moto";
+    protected static final String JSON_METADATA_KEY = "metadata";
+    protected static final String JSON_AUTH_MODE_KEY = "authorisation_mode";
+    protected static final String JSON_AUTH_MODE_MOTO_API = "moto_api";
 
     protected WorldpayMockClient worldpayMockClient;
     protected SmartpayMockClient smartpayMockClient;
